# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM Pay API",
        "description": "API de paiement mobile OM Pay avec authentification OTP",
        "contact": {
            "email": "support@ompay.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur de développement"
        },
        {
            "url": "https://annasock-ompay.onrender.com",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/{numero-compte}/solde": {
            "post": {
                "tags": [
                    "Comptes"
                ],
                "summary": "Consulter le solde d'un compte",
                "description": "Récupère le solde actuel d'un compte Orange Money. L'utilisateur doit être authentifié et propriétaire du compte.",
                "operationId": "getSolde",
                "parameters": [
                    {
                        "name": "numero-compte",
                        "in": "path",
                        "description": "Numéro du compte bancaire",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1234567890"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solde récupéré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde récupéré avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "compte_id": {
                                                    "type": "string",
                                                    "format": "uuid",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "solde": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500.5
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - L'utilisateur n'est pas propriétaire du compte",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/{numero_compte}/transfert": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Effectuer un transfert d'argent",
                "description": "Permet à un utilisateur authentifié d'effectuer un transfert d'argent vers un autre compte. L'utilisateur doit être propriétaire du compte émetteur.",
                "operationId": "transfert",
                "parameters": [
                    {
                        "name": "numero_compte",
                        "in": "path",
                        "description": "Numéro de compte de l'émetteur",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "1234567890"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_user_destinataire",
                                    "montant"
                                ],
                                "properties": {
                                    "numero_user_destinataire": {
                                        "description": "Numéro utilisateur du destinataire",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "771234568"
                                    },
                                    "montant": {
                                        "description": "Montant à transférer",
                                        "type": "number",
                                        "format": "float",
                                        "minimum": 0.01,
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert effectué avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solde insuffisant ou transfert à soi-même",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solde insuffisant"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Accès refusé - L'utilisateur n'est pas propriétaire du compte émetteur",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Accès non autorisé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Compte émetteur ou destinataire non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Compte non trouvé"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation des données d'entrée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your Bearer token in the format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}